requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
attrs==24.2.0
blinker==1.9.0
boto3==1.35.74
botocore==1.35.74
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
duckdb==1.1.3
Faker==19.13.0
fonttools==4.55.1
fpdf2==2.8.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
narwhals==1.15.2
numpy==1.26.4
openai==1.56.1
packaging==24.2
pandas==1.5.3
pandasai==2.4.0
pillow==10.4.0
protobuf==5.29.0
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.1
s3transfer==0.10.4
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlglot==25.32.1
sqlglotrs==0.3.0
streamlit==1.40.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==6.0.0

filelock~=3.16.1